+++ fe2o3-amqp/src/frames/error.rs
@@ -9,19 +9,18 @@ pub enum Error {
 
     /// AMQP error: decode error
     #[error("Decode Error")]
-    DecodeError,
+    DecodeError(String),
 
     /// AMQP error: not implemented
     #[error("AmqpError: NotImplemented")]
     NotImplemented,
 }
 
-// TODO: What about encode error?
 impl From<serde_amqp::Error> for Error {
     fn from(err: serde_amqp::Error) -> Self {
         match err {
             serde_amqp::Error::Io(e) => Self::Io(e),
-            _ => Self::DecodeError,
+            other => Self::DecodeError(other.to_string()),
         }
     }
 }