+++ fe2o3-amqp/Changelog.md
@@ -4,6 +4,7 @@
 
 1. ***Breaking*** changes:
    1. Restructured `connection::error::{OpenError, Error}` and `session::error:{BeginError, Error}`
+   2. `DecodeError` variant now carries a `String` message of the `serde_amqp::Error`
 2. `Connection` and non-txn `Session` no longer hold a copy of the controller sender to its own engine
 
 ## 0.3.2